+++ CHANGELOG.md
@@ -1,5 +1,6 @@
-### NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+### NEXT_VERSION_TYPE=PATCH
 ### NEXT_VERSION_DESCRIPTION_BEGIN
+* Исправлена конфигурация nexus-staging-plugin - теперь проставляется PackageGroup.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [7.3.0](https://github.com/yoomoney-gradle-plugins/gradle-project-plugin/pull/4) (20-04-2021)